# nest/docs/user-api-spec.yaml
openapi: 3.0.0
paths:
...
  /friendships/proposal:
    post:
      operationId: FriendshipsController_proposeFriendship
      summary: Send a friendship request
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposeFriendshipDto'
      responses:
        '201':
          description: ''
      tags: &ref_2
        - Friendships
      security: &ref_3
        - bearer: []
  /friendships/proposal/{targetId}:
    delete:
      operationId: FriendshipsController_cancelProposal
      summary: Cancel a friendship request
      parameters:
        - name: targetId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
  /friendships:
    get:
      operationId: FriendshipsController_getFriendships
      summary: Get friendships, either accepted ones, or sent/received requests
      parameters:
        - name: optionalFields[]
          required: true
          in: query
          schema:
            type: array
            items:
              type: string
        - name: type
          required: true
          in: query
          schema:
            enum:
              - proposed
              - received
              - established
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetFriendshipsResDto'
      tags: *ref_2
      security: *ref_3
  /friendships/request/{proposerId}:
    put:
      operationId: FriendshipsController_reviewRequest
      summary: Accept or reject a friendship request
      parameters:
        - name: proposerId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequestDto'
      responses:
        '200':
          description: ''
      tags: *ref_2
      security: *ref_3
...
  /interests:
    post:
      operationId: InterestsController_create
      summary: Create an interest relationship
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManualInterestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Interest'
      tags: &ref_6
        - Interests
      security: &ref_7
        - bearer: []
  /interests/shared:
    get:
      operationId: InterestsController_getShared
      summary: >-
        Get interests of a user that are shared with the ones of the
        authenticated user
      parameters:
        - name: userId
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSharedInterestsResDto'
      tags: *ref_6
      security: *ref_7
  /interests/me:
    get:
      operationId: InterestsController_getOwn
      summary: Get the interests of the authenticated user
      parameters:
        - name: resourceName
          required: false
          in: query
          schema:
            type: string
        - name: offset
          required: true
          in: query
          schema:
            type: number
        - name: limit
          required: true
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserInterestsResDto'
      tags: *ref_6
      security: *ref_7
  /interests/{id}:
    delete:
      operationId: InterestsController_delete
      summary: Delete an interest relationship
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_6
      security: *ref_7
  /texts:
    post:
      operationId: TextsController_create
      summary: Post a new text
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateManualTextDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Text'
      tags: &ref_8
        - Texts
      security: &ref_9
        - bearer: []
    get:
      operationId: TextsController_getOwn
      summary: Get the texts of the authenticated user
      parameters:
        - name: subtext
          required: false
          in: query
          schema:
            type: string
        - name: offset
          required: true
          in: query
          schema:
            type: number
        - name: limit
          required: true
          in: query
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserTextsResDto'
      tags: *ref_8
      security: *ref_9
  /texts/{id}:
    delete:
      operationId: TextsController_delete
      summary: Delete a text
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_8
      security: *ref_9
...
info:
  title: MindHub User API
  description: The user API to the MindHub system
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    ProfileDto:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
          enum:
            - man
            - woman
            - other
        age:
          type: number
        biography:
          type: string
      required:
        - name
        - gender
        - age
    ResourceDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - track
            - artist
      required:
        - id
        - name
        - type
    RelevanceDto:
      type: object
      properties:
        userId:
          type: string
        relevance:
          type: string
          enum:
            - normal
            - favorite
      required:
        - userId
        - relevance
    SharedInterestDto:
      type: object
      properties:
        resource:
          $ref: '#/components/schemas/ResourceDto'
        relevances:
          type: array
          items:
            $ref: '#/components/schemas/RelevanceDto'
      required:
        - resource
        - relevances
    UserPersonalityDto:
      type: object
      properties:
        userId:
          type: string
        o:
          type: number
        c:
          type: number
        e:
          type: number
        a:
          type: number
        'n':
          type: number
      required:
        - o
        - c
        - e
        - a
        - 'n'
    SharedUserInfo:
      type: object
      properties:
        _id:
          type: string
        profile:
          $ref: '#/components/schemas/ProfileDto'
        inactiveHours:
          type: number
        distance:
          type: number
        sharedInterests:
          type: array
          items:
            $ref: '#/components/schemas/SharedInterestDto'
        personality:
          $ref: '#/components/schemas/UserPersonalityDto'
        rating:
          type: number
      required:
        - _id
        - profile
        - inactiveHours
    ProposeFriendshipDto:
      type: object
      properties:
        target:
          type: string
      required:
        - target
    FriendshipScore:
      type: object
      properties:
        score:
          type: number
        by_ratings:
          type: number
        by_personality:
          type: number
      required:
        - score
        - by_ratings
        - by_personality
    CategoryInterestsScore:
      type: object
      properties:
        artist:
          type: number
        track:
          type: number
      required:
        - artist
        - track
    InterestsScore:
      type: object
      properties:
        score:
          type: number
        category:
          $ref: '#/components/schemas/CategoryInterestsScore'
      required:
        - score
        - category
    RecommendationScore:
      type: object
      properties:
        global:
          type: number
        friendship:
          $ref: '#/components/schemas/FriendshipScore'
        interests:
          $ref: '#/components/schemas/InterestsScore'
      required:
        - global
        - friendship
        - interests
    FriendshipDto:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/SharedUserInfo'
        score:
          $ref: '#/components/schemas/RecommendationScore'
      required:
        - user
        - score
    GetFriendshipsResDto:
      type: object
      properties:
        friends:
          type: array
          items:
            $ref: '#/components/schemas/FriendshipDto'
      required:
        - friends
    ReviewRequestDto:
      type: object
      properties:
        accept:
          type: boolean
      required:
        - accept
...
    CreateManualResourceDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - track
            - artist
      required:
        - name
        - type
    CreateManualInterestDto:
      type: object
      properties:
        relevance:
          type: string
          enum:
            - normal
            - favorite
        resource:
          $ref: '#/components/schemas/CreateManualResourceDto'
      required:
        - relevance
        - resource
    ObjectId:
      type: object
      properties: {}
    Resource:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - track
            - artist
      required:
        - id
        - name
        - type
    Interest:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        userId:
          type: string
        provider:
          type: string
          enum:
            - spotify
            - twitter
            - user
        relevance:
          type: string
          enum:
            - normal
            - favorite
        resource:
          $ref: '#/components/schemas/Resource'
        date:
          format: date-time
          type: string
      required:
        - userId
        - provider
        - relevance
        - resource
    GetSharedInterestsResDto:
      type: object
      properties:
        sharedInterests:
          type: array
          items:
            $ref: '#/components/schemas/SharedInterestDto'
      required:
        - sharedInterests
    UserInterestDto:
      type: object
      properties:
        _id:
          type: string
        provider:
          type: string
          enum:
            - spotify
            - twitter
            - user
        relevance:
          type: string
          enum:
            - normal
            - favorite
        resource:
          $ref: '#/components/schemas/ResourceDto'
        date:
          format: date-time
          type: string
      required:
        - _id
        - provider
        - relevance
        - resource
        - date
    GetUserInterestsResDto:
      type: object
      properties:
        interests:
          type: array
          items:
            $ref: '#/components/schemas/UserInterestDto'
        count:
          type: number
        total:
          type: number
      required:
        - interests
        - count
        - total
    CreateManualTextDto:
      type: object
      properties:
        rawText:
          type: string
        language:
          type: string
      required:
        - rawText
        - language
    Text:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        userId:
          type: string
        provider:
          type: string
          enum:
            - spotify
            - twitter
            - user
        rawText:
          type: string
        language:
          type: string
        date:
          format: date-time
          type: string
      required:
        - userId
        - provider
        - rawText
        - language
    UserTextDto:
      type: object
      properties:
        _id:
          type: string
        provider:
          type: string
          enum:
            - spotify
            - twitter
            - user
        rawText:
          type: string
        language:
          type: string
        date:
          format: date-time
          type: string
      required:
        - _id
        - provider
        - rawText
        - language
    GetUserTextsResDto:
      type: object
      properties:
        texts:
          type: array
          items:
            $ref: '#/components/schemas/UserTextDto'
        count:
          type: number
        total:
          type: number
      required:
        - texts
        - count
        - total
...